requirements.txt
django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
requests==2.28.2
gunicorn==20.1.0
pymssql==2.2.7